synchronize structure only of replicated relations
It's using the -t switch from pg_dump, thus with the same limitations.

At the moment we track only sequences and tables in pglogical, so this new
feature only dump/restore tables and linked objects (index, constraint,
sequences). Materialized view, for example, are not dump/restore at the moment.

For this first try, sequences are not explicitely retrieved, only those
corresponding to a replicated table. Eeven if the sequence itself is not subject
to replication...

More logic must be added for this feature to be complete.