Update protocol of write_rel
This is to allow target names to differ from the local names(we are on the
provider) when using write_rel).

New reltarget and nsptarget parameters are used as input instead of the Relation
information. This is because the target names can differ from the local names
(we are on the provider) when using write_rel).

NOTE:
if nsptarget/reltarget are set before the prepare plugin context function
then the regression_test "add_table" fails more often (because of timeout)...